+++ src/controllers/common/error.handler.js
@@ -6,7 +6,7 @@ export const errorHandler = ({ request, contentType, dataprocessed}) => {
   if (contentType.includes("application/json")) {
     const { responseType } = dataprocessed;
 
-    if (responseType !== 200) {
+    if (!!responseType && responseType !== 200) {
       store.dispatch({
         type: types.SET_ERROR,
         payload: dataprocessed,