+++ version-history.md
@@ -1,3 +1,4 @@
 |Versione|Descrizione|
 |-|-|
-|0.01.00|Presentazione a Mitrol dell'applicativo|
+|0.01.00|implementazione pagina login<br/>implementazione servizi api|
+|0.02.00|implementazione login-logout<br/>implementazione navigazione nell'applicazione<br/>gestione response model|